refactor: relocate shared suppression/policy loader to cli_params leaf (C7)
`_load_suppression_and_policy` is cross-command business logic (used by
`compare`, `compare-release`, `appcompat`, and the plugin command) but lived
inside the 2000-line `cli.py`, forcing three sibling CLI modules to import a
helper from the very file C7 is trying to shrink.

Move it next to `POLICY_FILE_PARAM` in the dependency-free `cli_params` leaf and
repoint every importer (`cli`, `cli_plugin`, `cli_compare_release`,
`cli_appcompat`) at the leaf. Pure relocation — no behaviour change. `cli.py`
drops 1993 → 1928 lines and its now-unused `PolicyFile`/`SuppressionList`
type-only imports.

Verified: full fast unit lane (10713 passed; the single failure,
test_collect_evidence_source_abi_graceful_without_tool, is environment-induced —
it asserts no source-ABI tool is installed and fails identically on HEAD without
this change once castxml is present locally; the CI unit lane installs no
castxml). ruff, mypy (0 errors), AI-readiness (0 errors, no new import cycle).
