+++ abicheck/cli_plugin.py
@@ -33,13 +33,12 @@ import click
 
 from .checker import Verdict
 from .cli import (
-    _load_suppression_and_policy,
     _resolve_input,
     _safe_write_output,
     _setup_verbosity,
     main,
 )
-from .cli_params import POLICY_FILE_PARAM
+from .cli_params import POLICY_FILE_PARAM, _load_suppression_and_policy
 
 
 def _load_required_entrypoints(
